pyproject.toml
[project]
name = "pythontemplate"
version = "0.1.0"
description = "My Python Template"
authors = [
    { name = "kaki005", email = "u097498f@ecs.osaka-u.ac.jp" }
]
dependencies = [
    "wandb>=0.17.6",
    "numpy>=2.0.1",
    "matplotlib>=3.9.1.post1",
    "hydra-core>=1.3.2",
    "scikit-learn>=1.5.1",
    "omegaconf>=2.3.0",
    "tqdm>=4.66.5",
    "seaborn>=0.13.2",
    "dataclasses-json>=0.6.7",
    "pandas>=2.2.2",
    "tensorly>=0.8.1",
    "pre-commit>=3.8.0",
    "opencv-python>=4.10.0.84",
    "opencv-contrib-python>=4.10.0.84",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/pythontemplate"]

[tool.ruff]
# 1行の最大文字数
line-length = 120

[tool.ruff.lint]
# チェックするエラーの種類
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
fixable = ["ALL"]
# 除外するエラーの種類
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs
    "B904", # raise ... from ... になっていない場合のエラーを無視
    "F401",   # imported but not used
    "UP008",
    "B007",
]

# ファイルごとのエラー除外
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

# isort の設定
[tool.ruff.lint.isort]
known-third-party = ["fastapi", "pydantic", "starlette"]

[tool.ruff.lint.pyupgrade]
# Python3.8互換のための設定
keep-runtime-typing = true



[tool.ruff.lint.mccabe]
max-complexity = 5


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
